move the server code into a subdirectory
