+++ src/server/registry.js
