+++ src/typescript-generator/specification.js
@@ -2,7 +2,7 @@ import nodePath from "node:path";
 
 import yaml from "js-yaml";
 
-import { readFile } from "../read-file.js";
+import { readFile } from "../util/read-file.js";
 
 import { Requirement } from "./requirement.js";
