+++ test/server/counterfact.test.js
@@ -1,9 +1,8 @@
 import supertest from "supertest";
 import Koa from "koa";
 
-import { counterfact } from "../src/counterfact.js";
-
-import { withTemporaryFiles } from "./lib/with-temporary-files.js";
+import { counterfact } from "../../src/server/counterfact.js";
+import { withTemporaryFiles } from "../lib/with-temporary-files.js";
 
 describe("integration test", () => {
   it("finds a path", async () => {
