+++ test/server/transpiler.test.js
@@ -2,9 +2,8 @@ import { once } from "node:events";
 import fs from "node:fs/promises";
 import { constants as fsConstants } from "node:fs";
 
-import { Transpiler } from "../src/transpiler.js";
-
-import { withTemporaryFiles } from "./lib/with-temporary-files.js";
+import { Transpiler } from "../../src/server/transpiler.js";
+import { withTemporaryFiles } from "../lib/with-temporary-files.js";
 
 const TYPESCRIPT_SOURCE = "const x:number = 1;\n";
 const JAVASCRIPT_SOURCE = "var x = 1;\n";
